+++ src/main/java/io/appform/dropwizard/actors/base/UnmanagedConsumer.java
@@ -88,6 +88,7 @@ public class UnmanagedConsumer<Message> {
                     log.info("Waiting for handler to complete processing the current message..");
                     while (handler.isRunning()) {
                         // wait for the handler to complete processing the current message
+                        Thread.onSpinWait();
                     }
                     channel.close();
                     log.info("Consumer channel closed for [{}] with prefix [{}]", name, config.getPrefix());