+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 ## 5.0.2-1
@@ -19,12 +20,15 @@ All notable changes to this project will be documented in this file.
   running for a queue.
 
 ## 2.0.28-12
+
 - Bug fixes and support for delayed in observer
 
 ## 2.0.28-11
+
 - Observers and operation level metrics support
 
 ## 2.0.28-10
+
 - Support custom replication factor in classic replicated queues
 - Support Lazy Queues
 - Support quorum queues with custom group size
@@ -33,19 +37,25 @@ All notable changes to this project will be documented in this file.
 - Optimize runtime of tests by introducing a singleton instance of RMQContainer
 
 ## 2.0.28-9
+
 Added header forwarding
 
 ## 2.0.28-2
+
 - Fixed pending count for sharded queues
 
 ## 1.3.18-2
+
 ### Added
+
 - BlockedListener for RMQ connections. Logs added to indicate connection blockage.
-- Support for separate producer/consumer connections. 
+- Support for separate producer/consumer connections.
 
 ### Removed
+
 - Removed `MetricRegistry` from RabbitMQBundle constructor. Registry is now picked from the supplied
 `Environment`
 
 ### Changed
+
 - Moved `ExecutorServiceProvider` from Bundle constructor to an overridden method. This allows access to config during `ExecutorServiceProvider` construction.